requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
aerich==0.7.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
itsdangerous==2.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

